+++ main.py
@@ -63,8 +63,14 @@ if user_choice == 'Add Knowledge':
         "Select Chunk Size", 100, 1000, st.session_state['chunk_size'], 50)
     st.session_state['chunk_overlap'] = st.sidebar.slider(
         "Select Chunk Overlap", 0, 100, st.session_state['chunk_overlap'], 10)
-    file_uploader(supabase, openai_api_key, vector_store)
-    url_uploader(supabase, openai_api_key, vector_store)
+    
+    # Create two columns for the file uploader and URL uploader
+    col1, col2 = st.columns(2)
+    
+    with col1:
+        file_uploader(supabase, openai_api_key, vector_store)
+    with col2:
+        url_uploader(supabase, openai_api_key, vector_store)
 elif user_choice == 'Chat with your Brain':
     # Display model and temperature selection only when asking questions
     st.sidebar.title("Configuration")
